future: fix wrong left menu link tooltip shown with the Popover
